+++ packages/python/qsu/verify/__init__.py
@@ -1,5 +1,6 @@
 from .between import between
 from .contains import contains
+from .hasBadWords import hasBadWords
 from .is2dArray import is2dArray
 from .isEmail import isEmail
 from .isEmpty import isEmpty
@@ -13,6 +14,7 @@ from .len import len
 __all__ = [
 	'between',
 	'contains',
+	'hasBadWords',
 	'is2dArray',
 	'isEmail',
 	'isEmpty',